disentangle feature flags
