+++ iroh-bytes/src/store.rs
@@ -1,8 +1,10 @@
 //! Implementations of blob stores
 use crate::{BlobFormat, Hash, HashAndFormat};
 
+#[cfg(feature = "file-db")]
 mod bao_file;
 pub mod mem;
+mod mutable_mem_storage;
 pub mod readonly_mem;
 
 #[cfg(feature = "file-db")]
